chore: rename GameSoundEffect and add index for Sound directory
