+++ frontend/src/pages/InGame/Sound/RoundSoundEffect.js
@@ -1,7 +1,7 @@
 import roundSound from '../../../assets/soundEffects/roundSuccess.mp3';
 
 // 효과음 재생 함수
-export const RoundSoundEffect = () => {
+const RoundSoundEffect = () => {
   const volume = 0.5;
   const audio = new Audio(roundSound);
   audio.volume = volume;
@@ -9,3 +9,5 @@ export const RoundSoundEffect = () => {
   // 사운드 재생
   audio.play();
 };
+
+export default RoundSoundEffect;
