+++ frontend/src/pages/InGame/Sound/index.js
@@ -1,0 +1,5 @@
+import BackgroundMusic from './BackgroundMusic';
+import MissionSoundEffects from './MissionSoundEffects';
+import RoundSoundEffect from './RoundSoundEffect';
+
+export { BackgroundMusic, MissionSoundEffects, RoundSoundEffect };
